mv to utils + fixes
